Create `userAttributes` folder to house functions to generate user attribute fields
